requirement.txt.txt
absl-py==0.11.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi @ file:///C:/ci/argon2-cffi_1596826740294/work
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
black==20.8b1
bleach==3.2.2
cachetools==4.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
cryptography==3.3.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
docutils==0.16
efficientnet==1.1.1
efficientnet-v2-keras @ git+https://github.com/sebastian-sz/efficientnet-v2-keras@94a2632c930318a4c6fd0fa03df0d5b1f5458e0f
entrypoints==0.3
flake8==3.8.4
focal-loss==0.0.7
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.35.0
h5py==2.10.0
helpdev==0.7.1
idna==2.10
image-classifiers==1.0.0
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.4.0
intervaltree==3.1.0
ipykernel==5.4.3
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core==4.7.0
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras==2.4.3
Keras-Applications==1.0.8
keras-efficientnet-v2==1.2.2
keras-efficientnets==0.1.7
keras-flops==0.1.2
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
notebook @ file:///C:/ci/notebook_1601498601160/work
numpy==1.18.5
numpydoc==1.1.0
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
patchify==0.2.2
pathspec==0.8.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.11
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.4
pylint==2.6.0
pyls-black==0.4.6
pyls-spyder==0.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
python-dateutil==2.8.1
python-git==2018.2.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2020.5
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==21.0.1
QDarkStyle==2.8.1
QtAwesome==1.0.2
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7
Rtree==0.9.7
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
segmentation-models @ git+https://github.com/qubvel/segmentation_models@94f624b7029deb463c859efbd92fa26f512b52b8
Send2Trash==1.5.0
sip==4.19.24
six==1.15.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==3.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
split-folders==0.4.3
spyder==4.2.1
spyder-kernels==1.10.1
statsmodels==0.12.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textdistance==4.2.0
threadpoolctl==2.1.0
three-merge==0.1.1
tifffile==2020.9.3
toml==0.10.2
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.2
watchdog==1.0.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
